handle Nunit v2 format
